часть рефакторинга
